backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Real-time event-driven monitoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.3.0",
    "bcrypt==4.0.1",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.48",
    "structlog>=25.5.0",
    "alembic>=1.18.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pylint>=4.0.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
